(genai) updated server schema format.
