--- genai/app/cache.py
@@ -1,49 +1,0 @@
-import redis
-import json
-from typing import Any, Optional
-import os
-from dotenv import load_dotenv
-
-load_dotenv()
-
-class CacheManager:
-    def __init__(self):
-        self.redis_client = redis.Redis(
-            host=os.getenv("REDIS_HOST", "localhost"),
-            port=int(os.getenv("REDIS_PORT", 6379)),
-            db=1,
-            decode_responses=True
-        ) if os.getenv("REDIS_HOST") else None
-    
-    def get(self, key: str) -> Optional[Any]:
-        """Get value from cache"""
-        if not self.redis_client:
-            return None
-        
-        try:
-            value = self.redis_client.get(key)
-            return json.loads(value) if value else None
-        except Exception:
-            return None
-    
-    def set(self, key: str, value: Any, expire: int = 3600) -> bool:
-        """Set value in cache with expiration"""
-        if not self.redis_client:
-            return False
-        
-        try:
-            self.redis_client.setex(key, expire, json.dumps(value, default=str))
-            return True
-        except Exception:
-            return False
-    
-    def delete(self, key: str) -> bool:
-        """Delete key from cache"""
-        if not self.redis_client:
-            return False
-        
-        try:
-            self.redis_client.delete(key)
-            return True
-        except Exception:
-            return False
